chore: format linting of jsx file
