+++ frontend/src/components/FormView.jsx
@@ -69,58 +69,44 @@ class FormView extends Component {
     };
 
     render() {
-        return ( <
-            div id = 'add-form' >
-            <
-            h2 > Add a New Trivia Question < /h2>  <
-            form className = 'form-view'
-            id = 'add-question-form'
-            onSubmit = { this.submitQuestion } >
-            <
-            label >
-            Question <
-            input type = 'text'
-            name = 'question'
-            onChange = { this.handleChange }
-            />  < /
-            label > <
-            label >
-            Answer <
-            input type = 'text'
-            name = 'answer'
-            onChange = { this.handleChange }
-            /> < /
-            label > <
-            label >
-            Difficulty <
-            select name = 'difficulty'
-            onChange = { this.handleChange } >
-            <
-            option value = '1' > 1 < /option> <
-            option value = '2' > 2 < /option> <
-            option value = '3' > 3 < /option> <
-            option value = '4' > 4 < /option> <
-            option value = '5' > 5 < /option> < /
-            select > <
-            /label> <
-            label >
-            Category <
-            select name = 'category'
-            onChange = { this.handleChange } > {
-                Object.keys(this.state.categories).map((id) => {
-                    return ( < option key = { id }
-                        value = { id } > { this.state.categories[id] } < /option>
-                    );
-                })
-            } <
-            /select> < /
-            label > <
-            input type = 'submit'
-            className = 'button'
-            value = 'Submit' / >
-            <
-            /form> < /
-            div >
+        return (
+            <div id = 'add-form' >
+                <h2 > Add a New Trivia Question </h2>
+                <form className = 'form-view'
+                id = 'add-question-form'
+                onSubmit = { this.submitQuestion } >
+                    <label>
+                        Question
+                        <input type = 'text'name = 'question' onChange = { this.handleChange }/>
+                    </label >
+                    <label>
+                        Answer
+                        <input type = 'text' name = 'answer' onChange = { this.handleChange }/>
+                    </label>
+                    <label>
+                        Difficulty
+                        <select name = 'difficulty' onChange = { this.handleChange }>
+                            <option value = '1' > 1 </option>
+                            <option value = '2' > 2 </option>
+                            <option value = '3' > 3 </option>
+                            <option value = '4' > 4 </option>
+                            <option value = '5' > 5 </option>
+                        </select >
+                    </label>
+                    <label>
+                        Category
+                        <select name = 'category' onChange = { this.handleChange } >
+                        {
+                        Object.keys(this.state.categories).map((id) => {
+                            return ( <option key = { id } value = { id } > { this.state.categories[id] }
+                            </option>);
+                            })
+                            }
+                        </select>
+                    </label>
+                <input type = 'submit' className = 'button' value = 'Submit' />
+            </form>
+        </div>
         );
     }
 }
